requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
decouple==0.0.7
Django==3.2.21
django-auth-token==0.2.15
django-chamber==0.6.17
django-choice-enumfields==1.1.3
django-generic-m2m-field==0.0.9
django-ipware==5.0.0
django-restframework==0.0.1
djangorestframework==3.14.0
filemagic==1.6
idna==3.4
import-string==0.1.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
openapi-codec==1.3.2
Pillow==10.0.1
PyPrind==2.11.3
pytz==2023.3.post1
requests==2.31.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.8.0
Unidecode==1.3.6
uritemplate==4.1.1
urllib3==2.0.5
